requirements.txt
google-generativeai
pillow
opencv-python
pyttsx3
gTTS
pyobjc
python-dotenv
SpeechRecognition
PyAudio